@@ -3,13 +3,13 @@
 	"description": "Synchronise two directories.",
 
 	"require": {
-		"php": ">=7.2",
+		"php": ">=8.0",
 		"phpgt/cli": "1.3.*",
 		"webmozart/glob": "4.*"
 	},
 
 	"require-dev": {
-		"phpunit/phpunit": "9.*",
+		"phpunit/phpunit": "^9.5",
 		"phpstan/phpstan": "^0.12"
 	},
 
